@@ -7,20 +7,22 @@ import 'package:http/http.dart' as http;
 
 class AppStateModel extends ChangeNotifier {
   AppStateModel({
-    @required this.allCommoditiesHistory,
-    @required this.interestedInPrices,
+    required this.allCommoditiesHistory,
+    required this.interestedInPrices,
   });
 
   Map<String, List<PriceCheck>> allCommoditiesHistory;
   Map<String, double> interestedInPrices;
+  // TODO change to minutes
+  int refreshRateInSeconds = 10;
 
   AppStateModel copyWith({
-    Map<String, List<PriceCheck>> allCommoditiesHistory,
-    Map<String, double> interestedInPrices,
+    Map<String, List<PriceCheck>>? allCommoditiesHistory,
+    Map<String, double>? interestedInPrices,
   }) {
     var _newAppState = AppStateModel(
       allCommoditiesHistory:
-      allCommoditiesHistory ?? this.allCommoditiesHistory,
+          allCommoditiesHistory ?? this.allCommoditiesHistory,
       interestedInPrices: interestedInPrices ?? this.interestedInPrices,
     );
     return _newAppState;
@@ -28,63 +30,63 @@ class AppStateModel extends ChangeNotifier {
 
   AppStateModel.initialState()
       : allCommoditiesHistory = {
-    'AAVE-USD': [],
-    'AAVE-USDT': [],
-    'ALGO-BTC': [],
-    'ALGO-USD': [],
-    'ALGO-USDT': [],
-    'BCH-BTC': [],
-    'BCH-ETH': [],
-    'BCH-EUR': [],
-    'BCH-PAX': [],
-    'BCH-USD': [],
-    'BCH-USDT': [],
-    'BTC-EUR': [],
-    'BTC-GBP': [],
-    'BTC-PAX': [],
-    'BTC-TRY': [],
-    'BTC-USD': [],
-    'BTC-USDT': [],
-    'DGLD-BTC': [],
-    'DGLD-USD': [],
-    'ENJ-USD': [],
-    'ENJ-USDT': [],
-    'ETH-BTC': [],
-    'ETH-EUR': [],
-    'ETH-GBP': [],
-    'ETH-PAX': [],
-    'ETH-TRY': [],
-    'ETH-USD': [],
-    'ETH-USDT': [],
-    'LEND-USD': [],
-    'LEND-USDT': [],
-    'LTC-BTC': [],
-    'LTC-EUR': [],
-    'LTC-PAX': [],
-    'LTC-TRY': [],
-    'LTC-USD': [],
-    'LTC-USDT': [],
-    'OGN-USD': [],
-    'OGN-USDT': [],
-    'PAX-EUR': [],
-    'PAX-USD': [],
-    'USDT-EUR': [],
-    'USDT-GBP': [],
-    'USDT-TRY': [],
-    'USDT-USD': [],
-    'WDGLD-BTC': [],
-    'WDGLD-DGLD': [],
-    'WDGLD-USD': [],
-    'XLM-BTC': [],
-    'XLM-ETH': [],
-    'XLM-EUR': [],
-    'XLM-PAX': [],
-    'XLM-USD': [],
-    'XRP-EUR': [],
-    'XRP-USD': [],
-    'YFI-USD': [],
-    'YFI-USDT': [],
-  },
+          'AAVE-USD': [],
+          'AAVE-USDT': [],
+          'ALGO-BTC': [],
+          'ALGO-USD': [],
+          'ALGO-USDT': [],
+          'BCH-BTC': [],
+          'BCH-ETH': [],
+          'BCH-EUR': [],
+          'BCH-PAX': [],
+          'BCH-USD': [],
+          'BCH-USDT': [],
+          'BTC-EUR': [],
+          'BTC-GBP': [],
+          'BTC-PAX': [],
+          'BTC-TRY': [],
+          'BTC-USD': [],
+          'BTC-USDT': [],
+          'DGLD-BTC': [],
+          'DGLD-USD': [],
+          'ENJ-USD': [],
+          'ENJ-USDT': [],
+          'ETH-BTC': [],
+          'ETH-EUR': [],
+          'ETH-GBP': [],
+          'ETH-PAX': [],
+          'ETH-TRY': [],
+          'ETH-USD': [],
+          'ETH-USDT': [],
+          'LEND-USD': [],
+          'LEND-USDT': [],
+          'LTC-BTC': [],
+          'LTC-EUR': [],
+          'LTC-PAX': [],
+          'LTC-TRY': [],
+          'LTC-USD': [],
+          'LTC-USDT': [],
+          'OGN-USD': [],
+          'OGN-USDT': [],
+          'PAX-EUR': [],
+          'PAX-USD': [],
+          'USDT-EUR': [],
+          'USDT-GBP': [],
+          'USDT-TRY': [],
+          'USDT-USD': [],
+          'WDGLD-BTC': [],
+          'WDGLD-DGLD': [],
+          'WDGLD-USD': [],
+          'XLM-BTC': [],
+          'XLM-ETH': [],
+          'XLM-EUR': [],
+          'XLM-PAX': [],
+          'XLM-USD': [],
+          'XRP-EUR': [],
+          'XRP-USD': [],
+          'YFI-USD': [],
+          'YFI-USDT': [],
+        },
         interestedInPrices = <String, double>{
           'BTC-USD': 100,
           'ETH-USD': 50,
@@ -100,11 +102,11 @@ class AppStateModel extends ChangeNotifier {
   // ;
 
   Map<String, dynamic> toJson() => {
-    'allCommoditiesHistory': allCommoditiesHistory,
-  };
+        'allCommoditiesHistory': allCommoditiesHistory,
+      };
 
   // Returns a list of the latest PriceCheck objects
-  Future<List<dynamic>>  getTicker() async {
+  Future<List<dynamic>> getTicker() async {
     List<dynamic> _jsonData = [];
     List<dynamic> _data = [];
 
@@ -132,10 +134,11 @@ class AppStateModel extends ChangeNotifier {
   }
 
   void addUpdatesToHistory({
-    @required List<PriceCheck> newestUpdates,
-    @required Map<String, List<PriceCheck>> outdatedHistory,
+    required List<dynamic> newestUpdates,
+    required Map<String, List<PriceCheck>> outdatedHistory,
   }) {
-    var _replacementHistory = Map<String, List<PriceCheck>>.from(outdatedHistory);
+    var _replacementHistory =
+        Map<String, List<PriceCheck>>.from(outdatedHistory);
     // Add each new listing to the result
     for (int i = 0; i < newestUpdates.length; i++) {
       var temp = PriceCheck(
@@ -144,28 +147,27 @@ class AppStateModel extends ChangeNotifier {
         price24h: newestUpdates[i].price24h,
         volume24h: newestUpdates[i].volume24h,
       );
-      _replacementHistory['${temp.symbol}'].add(temp);
+      _replacementHistory['${temp.symbol}']!.add(temp);
     }
-    allCommoditiesHistory = Map<String, List<PriceCheck>>.from(_replacementHistory);
+    allCommoditiesHistory =
+        Map<String, List<PriceCheck>>.from(_replacementHistory);
 
     Map<String, double> _replacementWatchList = interestedInPrices;
     for (String key in interestedInPrices.keys) {
-      _replacementWatchList[key] = _replacementHistory[key].last.lastTradePrice;
+      _replacementWatchList[key] = _replacementHistory[key]!.last.lastTradePrice;
     }
     interestedInPrices = _replacementWatchList;
     notifyListeners();
   }
 
-  Future<List<dynamic>> fetchAndProcessUpdates(
-      {
-        @required Map<String, List<PriceCheck>> outdatedHistory}) async {
-    List<dynamic> _result;
+  Future<void> fetchAndProcessUpdates(
+      {required Map<String, List<PriceCheck>> outdatedHistory}) async {
     await getTicker().then((value) {
       addUpdatesToHistory(
         newestUpdates: value,
         outdatedHistory: outdatedHistory,
       );
     });
-    return _result;
+
   }
 }
