@@ -1,21 +1,36 @@
 import 'package:crypto_tracker_redux/app/app_strings.dart';
+import 'package:crypto_tracker_redux/app/app_textstyles.dart';
 import 'package:crypto_tracker_redux/provider_version/models/app_state_model.dart';
-import 'package:crypto_tracker_redux/provider_version/models/price_check_model.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-class ProviderHome extends StatelessWidget {
-  const ProviderHome({
-    Key key,
+class ProviderHome extends StatefulWidget {
+
+  ProviderHome({
+    Key? key,
   }) : super(key: key);
 
   @override
+  _ProviderHomeState createState() => _ProviderHomeState();
+}
+
+class _ProviderHomeState extends State<ProviderHome> {
+
+  @override
+  void initState() {
+    super.initState();
+  }
+
+
+
+  @override
   Widget build(BuildContext context) {
-    final Map<String, List<PriceCheck>> allCommoditiesHistory =
-        Provider.of<AppStateModel>(context).allCommoditiesHistory;
-    final Map<String, double> interestedInPrices =
-        Provider.of<AppStateModel>(context).interestedInPrices;
-    return Center(
+    // Fetch the values we care about
+    final interestedInPrices = context.select((AppStateModel appState) => appState.interestedInPrices);
+    final allCommoditiesHistory = context.select((AppStateModel appState) => appState.allCommoditiesHistory);
+    //
+    return Padding(
+      padding: const EdgeInsets.all(16),
       child: Column(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         crossAxisAlignment: CrossAxisAlignment.center,
@@ -30,7 +45,7 @@ class ProviderHome extends StatelessWidget {
                     String value = interestedInPrices[key].toString();
                     final regex = RegExp(r'^([A-z]+)-([A-z]+)$');
                     final regexMatch = regex.firstMatch(key);
-                    final currencyInterestedIn = regexMatch.group(1);
+                    final currencyInterestedIn = regexMatch!.group(1);
                     final denomination = regexMatch.group(2);
                     return Card(
                       borderOnForeground: true,
@@ -41,13 +56,10 @@ class ProviderHome extends StatelessWidget {
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: <Widget>[
                             Text(
-                              '$currencyInterestedIn in $denomination is',
-                              style: Theme.of(context).textTheme.headline4,
+                              '${appStrings[currencyInterestedIn]} is currently \n${value.padRight(2)} ${appStrings[denomination]}',
+                              textAlign: TextAlign.center,
+                              style: appTextStyles['normal24'],
                             ),
-                            Text(
-                              '$value',
-                              style: Theme.of(context).textTheme.headline4,
-                            )
                           ],
                         ),
                       ),
@@ -57,24 +69,29 @@ class ProviderHome extends StatelessWidget {
               );
             },
           ),
-          ElevatedButton(
-            onPressed: () => Provider.of<AppStateModel>(context, listen: false)
-                .fetchAndProcessUpdates(
-              outdatedHistory: allCommoditiesHistory,
-            ),
-            style: ButtonStyle(
-              backgroundColor: MaterialStateProperty.resolveWith<Color>(
-                (Set<MaterialState> states) {
-                  if (states.contains(MaterialState.pressed))
-                    return Theme.of(context)
-                        .colorScheme
-                        .primary
-                        .withOpacity(0.5);
-                  return null; // Use the component's default.
-                },
+          Row(
+            children: [
+              ElevatedButton(
+                onPressed: () => Provider.of<AppStateModel>(context, listen: false)
+                    .fetchAndProcessUpdates(
+                  outdatedHistory: allCommoditiesHistory,
+                ),
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.resolveWith<Color?>(
+                    (Set<MaterialState> states) {
+                      if (states.contains(MaterialState.pressed))
+                        return Theme.of(context)
+                            .colorScheme
+                            .primary
+                            .withOpacity(0.5);
+                      return null; // Use the component's default.
+                    },
+                  ),
+                ),
+                child: Text('Update Prices'),
               ),
-            ),
-            child: Text('Update Prices'),
+
+            ],
           ),
         ],
       ),
