@@ -1,40 +1,39 @@
 import 'package:flutter/material.dart';
 
-class AppColors {
-
+const Map<String, Color> appColors = {
   // Custom Colors
-  static const Color loginAndSignupBoxes = Color(0x55768fff);
-  static const Color loginAndSignupBoxesBorder = Color(0x77000000);
-  static const Color customDropShadowColor = Color(0x99000000);
+  'loginAndSignupBoxes': Color(0x55768fff),
+  'loginAndSignupBoxesBorder': Color(0x77000000),
+  'customDropShadowColor': Color(0x99000000),
 
   // General Colors
-  static const Color whiteTextColor = Color(0xDDFFFFFF);
-  static const Color blackTextColor = Color(0xFF111111);
-  static const Color dropShadowColor = Color(0x33000000);
-  static const Color offWhitePageBackground = Color(0xFFF9F9F9);
+  'whiteTextColor': Color(0xDDFFFFFF),
+  'blackTextColor': Color(0xFF111111),
+  'dropShadowColor': Color(0x33000000),
+  'offWhitePageBackground': Color(0xFFF9F9F9),
 
-  static const Color bronze = Color(0xFFD6943A);
-  static const Color silver = Color(0xFFD9D9D9);
-  static const Color gold = Color(0xFFF7D44E);
+  'bronze': Color(0xFFD6943A),
+  'silver': Color(0xFFD9D9D9),
+  'gold': Color(0xFFF7D44E),
 
   // Material Design Color Sets
-  static const Color primaryBlue = Color(0xff2962ff);
-  static const Color primaryLightBlue = Color(0xff768fff);
-  static const Color primaryDarkBlue = Color(0xff0039cb);
+  'primaryBlue': Color(0xff2962ff),
+  'primaryLightBlue': Color(0xff768fff),
+  'primaryDarkBlue': Color(0xff0039cb),
 
-  static const Color primaryRed = Color(0xffd50000);
-  static const Color primaryLightRed = Color(0xffff5131);
-  static const Color primaryDarkRed = Color(0xff9b0000);
+  'primaryRed': Color(0xffd50000),
+  'primaryLightRed': Color(0xffff5131),
+  'primaryDarkRed': Color(0xff9b0000),
 
-  static const Color primaryGreen = Color(0xff00c853);
-  static const Color primaryLightGreen = Color(0xff5efc82);
-  static const Color primaryDarkGreen = Color(0xff009624);
+  'primaryGreen': Color(0xff00c853),
+  'primaryLightGreen': Color(0xff5efc82),
+  'primaryDarkGreen': Color(0xff009624),
 
-  static const Color primaryPurple = Color(0xffaa00ff);
-  static const Color primaryLightPurple = Color(0xffe254ff);
-  static const Color primaryDarkPurple = Color(0xff7200ca);
+  'primaryPurple': Color(0xffaa00ff),
+  'primaryLightPurple': Color(0xffe254ff),
+  'primaryDarkPurple': Color(0xff7200ca),
 
-  static const Color primaryDarkYellow = Color(0xffffd600);
-  static const Color primaryDarkLightYellow = Color(0xffffff52);
-  static const Color primaryDarkDarkYellow = Color(0xffc7a500);
-}
+  'primaryDarkYellow': Color(0xffffd600),
+  'primaryDarkLightYellow': Color(0xffffff52),
+  'primaryDarkDarkYellow': Color(0xffc7a500),
+};
