@@ -1,176 +1,3 @@
-class AppImages {
-  static List<String> availableImages = [
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    //
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-    bitInComforter,
-    bitInBackyard,
-    bitOnDanceTrack,
-    bitBelly,
-    bitChillin,
-
-  ];
-  static const String bitInComforter = 'assets/images/2009-07-02.jpg';
-  static const String bitInBackyard = 'assets/images/2009-07-09.jpg';
-  static const String bitOnDanceTrack = 'assets/images/1231345588676.jpg';
-  static const String bitBelly = 'assets/images/bit_belly.jpg';
-  static const String bitChillin = 'assets/images/Bit_chilling.jpg';
-}
+  const Map<String, String> appImages  = {
+  //  'sampleImageListing' : 'assets/images/XXXXXXXXXXXXXX.jpg',
+  };
