@@ -1,5 +1,8 @@
 import 'package:crypto_tracker_redux/pages/home.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import 'provider_version/models/app_state_model.dart';
 
 void main() {
   runApp(MyApp());
@@ -12,10 +15,13 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return  MaterialApp(
-        title: 'Redux ToDo Demo',
-        theme: ThemeData.dark(),
-        home: const Home(),
+    return MaterialApp(
+      title: 'Redux ToDo Demo',
+      theme: ThemeData.dark(),
+      home: ChangeNotifierProvider(
+        create: (BuildContext context) => AppStateModel.initialState(),
+        child: Home(),
+      ),
     );
   }
 }
